+++ FaceAnalyzer/__init__.py
@@ -14,4 +14,4 @@
 <================"""
 from .Face import Face, DrawingSpec
 from .FaceAnalyzer import FaceAnalyzer
-from .Helpers import buildCameraMatrix, faceOrientation2Euler
+from .Helpers import buildCameraMatrix, faceOrientation2Euler, get_z_line_equation